EPBDS-13012 Refactoring. Merge two classes
